Rework firmware to make more sense on the host end.
It turns out that the board with a MAX7219 and an 8x8 LED matrix I was
using for testing has a funky pinout which doesn't match how the Panda
Display is wired.  Pin 1 does not correspond to column 1 and so on. Then
I discovered that I wired the board such that I need to rotate the
bitmap 90 degrees clockwise.  I could possibly rework the board design
to save time in the microcontroller.  Maybe it's not necessary,
particularly given the thing is running at 16MHz.